:recycle: refactor: 테스트 코드 준비
