--- backend/src/test/java/kernel360/techpick/folder/service/parser/DummyNameProvider.java
@@ -1,16 +1,0 @@
-package kernel360.techpick.folder.service.parser;
-
-import kernel360.techpick.feature.structure.model.NameProvider;
-
-public class DummyNameProvider implements NameProvider {
-
-	@Override
-	public String findPickNameById(Long id) {
-		return "DUMMY_PICK_NAME";
-	}
-
-	@Override
-	public String findFolderNameById(Long id) {
-		return "DUMMY_FOLDER_NAME";
-	}
-}
